Length

+++ construct.cpp
@@ -156,7 +156,7 @@ inline void run(Options & options)
     double start = sysTime();
     construct(index);
     double finish = sysTime();
-    std::cout << length(text) << " texts" << std::endl;
+    std::cout << (unsigned long)length(text) << " texts" << std::endl;
     std::cout << lengthSum(text) << " symbols" << std::endl;
     std::cout << finish - start << " sec" << std::endl;
 //    std::cout << (unsigned)(lengthSum(text) / (finish - start)) << " symbols/sec" << std::endl;
